최종0324

+++ src/main/webapp/shop_db/deleteCartSel.jsp
@@ -1,6 +1,7 @@
 <%@ page language="java" contentType="text/html; charset=UTF-8"
          pageEncoding="UTF-8"%>
 <%@ page import="com.teamProject.cart.model.CartDAO" %>
+<%@ page import="java.io.PrintWriter" %>
 
 <%
 
@@ -10,7 +11,15 @@
 
   // 문자열로 넘어온 chkdID 사용.
   String chkdID = request.getParameter("chkdID");
-  cartDAO.deleteCartSel(orderNum, chkdID);
+
+  PrintWriter write = response.getWriter();
+  if (chkdID == "") {
+    write.println("<script>alert('삭제할 상품을 선택해 주세요'); location.href='cart.jsp';</script>");
+    write.flush();
+  }
+  else {
+    cartDAO.deleteCartSel(orderNum, chkdID);
+  }
 
   response.sendRedirect("cart.jsp");
 %>
